+++ analysis_options.yaml
@@ -14,6 +14,7 @@ linter:
     - prefer_const_constructors
     - prefer_const_literals_to_create_immutables
     - unawaited_futures
+    - migrate_design_widgets
 
 formatter:
   trailing_commas: preserve
