+++ lib/pages/settings/settings_page.dart
@@ -5,7 +5,7 @@ import 'package:kover/pages/settings/credentials_settings.dart';
 import 'package:kover/pages/settings/data_management_settings.dart';
 import 'package:kover/pages/settings/general_settings.dart';
 import 'package:kover/pages/settings/version_label.dart';
-import 'package:kover/widgets/util/breakpoint_builder.dart';
+import 'package:kover/widgets/util/sliver_adaptive_padding.dart';
 import 'package:kover/widgets/util/sliver_bottom_padding.dart';
 import 'package:material_ui/material_ui.dart';
